+++ backend/package.json
@@ -20,12 +20,15 @@
     "typescript": "5.4.2"
   },
   "dependencies": {
+    "axios": "^1.6.8",
     "express": "4.18.3"
   },
   "eslintConfig": {
     "parserOptions": {
       "project": "./tsconfig.json"
     },
-    "extends": ["./node_modules/ts-standard/eslintrc.json"]
+    "extends": [
+      "./node_modules/ts-standard/eslintrc.json"
+    ]
   }
 }
